Kernel+Userland: Make some of the POSIX types larger
Expand the following types from 32-bit to 64-bit:
- blkcnt_t
- blksize_t
- dev_t
- nlink_t
- suseconds_t
- clock_t

This matches their size on other 64-bit systems.

+++ Kernel/API/POSIX/sys/types.h
@@ -38,16 +38,16 @@ typedef int id_t;
 typedef uint64_t ino_t;
 typedef int64_t off_t;
 
-typedef uint32_t blkcnt_t;
-typedef uint32_t blksize_t;
-typedef uint32_t dev_t;
+typedef uint64_t blkcnt_t;
+typedef uint64_t blksize_t;
+typedef uint64_t dev_t;
 typedef uint16_t mode_t;
-typedef uint32_t nlink_t;
+typedef uint64_t nlink_t;
 
 typedef int64_t time_t;
 typedef uint32_t useconds_t;
-typedef int32_t suseconds_t;
-typedef uint32_t clock_t;
+typedef int64_t suseconds_t;
+typedef uint64_t clock_t;
 
 typedef uint64_t fsblkcnt_t;
 typedef uint64_t fsfilcnt_t;
